lineage/adf-copy-activity/src/azfunctions/requirements.txt
azure-functions
msrest
pyodbc